pyproject.toml
[project]
name = "pampu"
version = "0.0.0.dev0"
description = "CLI for Atlassian Bamboo"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["bamboo", "cli", "atlassian", "ci", "cd"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "atlassian-python-api>=3.41.0",
    "platformdirs>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/vivainio/pampu"
Repository = "https://github.com/vivainio/pampu"
Issues = "https://github.com/vivainio/pampu/issues"

[project.scripts]
pampu = "pampu.cli:main"

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.uv]
package = true

[tool.setuptools]
packages = ["pampu"]

[tool.ruff]
line-length = 100
